Change IMMORTAL_CHANCE config type to Double

+++ common/src/main/java/com/github/teamfusion/rottencreatures/ConfigEntries.java
@@ -11,5 +11,5 @@ public class ConfigEntries {
     public static final Config.Entry<Integer> UNDEAD_MINER_DEPTH = Config.create("Undead Miner Depth", 63, "determines the max y height where Undead Miners can spawn");
     public static final Config.Entry<Integer> MUMMY_WEIGHT = Config.create("Mummy Weight", 20, "determines how often will Mummies spawn");
     public static final Config.Entry<Integer> DEAD_BEARD_WEIGHT = Config.create("Dead Beard Weight", 1, "determines how often will Dead Beard spawn");
-    public static final Config.Entry<Float> IMMORTAL_CHANCE = Config.create("Immortal Chance", 0.025F, "determines the chance of spawning for Immortals");
+    public static final Config.Entry<Double> IMMORTAL_CHANCE = Config.create("Immortal Chance", 0.025D, "determines the chance of spawning for Immortals");
 }
